+++ package.json
@@ -50,6 +50,7 @@
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.7",
+    "jsdom": "^26.1.0",
     "prettier": "^3.6.2",
     "tsx": "^4.17.0",
     "typescript": "^5.5.4",
